Add modules task
